Remove any keyword from the SMS before forwarding to RapidPro message.
Sometimes Kannel servers are configured to forward and send a message to
us only if it starts with a specific keyword. But this keyword isn't what
RapidPro understands and uses as a trigger. Hence, we remove this keyword
from the message,if present in the beginning, before forwarding to RapidPro
so that RapidPro recognizes and processes the right keyword (the second word
in the original SMS)